NBench v0.2 stable release

+++ src/NBench/Collection/Memory/GcTotalMemoryCollector.cs
@@ -3,6 +3,7 @@
 
 using System;
 using NBench.Metrics;
+using NBench.Metrics.Memory;
 
 namespace NBench.Collection.Memory
 {
@@ -23,7 +24,7 @@ namespace NBench.Collection.Memory
         {
         }
 
-        public override long Collect()
+        public override double Collect()
         {
             /*
              * We intentionally don't allow the garbage collector to collect anything
